Implemented continuations in the context of a multi-processing environment. Now continuations can jump across processes. When one process invokes a continuation that was captured in another process, the execution jumps to the other process first, but the original process is not destroyed, but the continuation invocation will return with MUSE_NIL in the calling process so both processes proceed.
Continuation capturing has also been rendered more
efficient as a consequence of the process implementation. Now all that the capturing operation has to do is to 
copy the stacks. There is no stepping through the list
of symbols to save all the bindings. The memory 
consumed is therefore lower as well since only the
values need to be saved, not the symbols themselves
Previously, the bindings were saved as an array of
symbol-value pairs, but now they are just an array
of process-local-values of each symbol.