Issue #20 - Some compatibility with standard scheme idioms will be helpful in porting.
Call a lambda a lambda! muSE uses the symbol "fn" to mean what is referred to as "lambda" in standard Scheme. This is 'cos muSE is used by people who might be scared away if they see too many "lambda" expressions floating around ... but they understand functions (thanks to JavaScript) and can easily read "fn" as an abbreviation of "function". 

That said, this breaks code compatibility. So now, with this change, both "fn" and "lambda" can be used.