FEATURE: Made module names behave like "do", except that within the scope of the module expression's body, symbols are interpreted in the scope of the module - irrespective of whether the module expression is at the global level or within a closure creating expression.
This feature is made possible by the introduction of a 
new "scope view" - id 'scop' - which lets an object provide
a hook into the closure creation code to introduce local
bindings that aren't visible in the s-expression itself. 

The scope view lets us define context sensitive forms.
Ultimately, all binding introducing forms -
"define", "let", "case", "fn", "fn:" and "gfn" -
can be implemented by providing a different scope
view for each so that there is a constant testing overhead
during closure creation instead of depending on the number
of such special forms that need to be supported.